lp64: checkpoint train-flow hardening before batch sweeps
Capture current pointer, UB, and watchdog fixes plus fast boot hooks so we can branch into broader LP64 sweeps from a known-good baseline.

+++ include/dolphin/os/OSUtil.h
@@ -8,8 +8,8 @@ extern "C"
 {
 #endif
 
-#define OSRoundUp32B(x) (((u32)(x) + 0x1F) & ~(0x1F))
-#define OSRoundDown32B(x) (((u32)(x)) & ~(0x1F))
+#define OSRoundUp32B(x) ((((uintptr_t)(x)) + (uintptr_t)0x1F) & ~(uintptr_t)0x1F)
+#define OSRoundDown32B(x) (((uintptr_t)(x)) & ~(uintptr_t)0x1F)
 
 #ifdef __cplusplus
 };
